requirements.txt
# Core frameworks
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0

# Data processing
numpy>=1.24.0
pandas>=2.0.0
pillow>=9.5.0
scikit-learn>=1.2.0

# Visualization (optional, if you use it)
matplotlib>=3.7.0